feat(S11P21A705-625): 주식 환율 추가

+++ backend/src/main/java/com/zzf/backend/domain/animal/dto/AnimalInfoResponse.java
@@ -13,7 +13,7 @@ public record AnimalInfoResponse(String animalName,
                                  Long totalAssets,
                                  Long totalDeposit,
                                  Long totalSavings,
-                                 Double totalStock,
+                                 Long totalStock,
                                  Long totalLoan,
                                  Long totalCapital) {
 }
